fix: keep tiled builders side-effect-free
Split the tiled box and plane builders into pure implementations and compatibility wrappers so Meshes/pure remains side-effect-free while legacy VertexData registrations still run from non-pure entrypoints.
