+++ packages/dev/core/src/Meshes/meshBuilder.ts
@@ -3,6 +3,7 @@ export * from "./meshBuilder.pure";
 import "./Builders/ribbonBuilder";
 import "./Builders/discBuilder";
 import "./Builders/boxBuilder";
+import "./Builders/tiledBoxBuilder";
 import "./Builders/sphereBuilder";
 import "./Builders/cylinderBuilder";
 import "./Builders/torusBuilder";
@@ -12,6 +13,7 @@ import "./Builders/polygonBuilder";
 import "./Builders/shapeBuilder";
 import "./Builders/latheBuilder";
 import "./Builders/planeBuilder";
+import "./Builders/tiledPlaneBuilder";
 import "./Builders/groundBuilder";
 import "./Builders/tubeBuilder";
 import "./Builders/polyhedronBuilder";
