+++ packages/dev/core/test/unit/Meshes/meshBuilder.sideEffects.test.ts
@@ -2,6 +2,7 @@ import { describe, expect, it } from "vitest";
 
 import { NullEngine } from "core/Engines/nullEngine";
 import { Mesh } from "core/Meshes/mesh.pure";
+import { VertexData } from "core/Meshes/mesh.vertexData";
 import { Scene } from "core/scene";
 import "core/Meshes/meshBuilder";
 
@@ -19,5 +20,10 @@ describe("MeshBuilder side effects", () => {
             scene.dispose();
             engine.dispose();
         }
+    });
+
+    it("registers legacy tiled VertexData creation functions", () => {
+        expect(() => VertexData.CreateTiledPlane({})).not.toThrow();
+        expect(() => VertexData.CreateTiledBox({})).not.toThrow();
     });
 });
