+++ packages/public/@babylonjs/core/package.json
@@ -454,6 +454,8 @@
         "Meshes/Builders/polyhedronBuilder.js",
         "Meshes/Builders/ribbonBuilder.js",
         "Meshes/Builders/shapeBuilder.js",
+        "Meshes/Builders/tiledBoxBuilder.js",
+        "Meshes/Builders/tiledPlaneBuilder.js",
         "Meshes/Builders/torusBuilder.js",
         "Meshes/Builders/torusKnotBuilder.js",
         "Meshes/Builders/tubeBuilder.js",
